requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
transformers==4.44.2
peft==0.12.0
torch==2.4.1
python-dotenv==1.0.1
pydantic==2.9.2